% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgnbd.R
\name{bgnbd}
\alias{bgnbd}
\title{Fit a BG/NBD model with or without random effects}
\usage{
bgnbd(
  rfmt,
  reff = c("marginal", "prob", "lambda", "full"),
  pars_alpha = c(0.01, 0.01),
  pars_r = c(0.01, 0.01),
  pars_a = c(0.01, 0.01),
  pars_b = c(0.01, 0.01),
  approach = c("bayesian", "classical"),
  gen_loglik = c("no", "yes"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Fit a BG/NBD model with or without random effects
}
