% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bytd-package.R
\docType{package}
\name{bytd-package}
\alias{bytd-package}
\alias{bytd}
\title{bytd: Bayesian implementation of BG/NBD for CLV modeling}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
