% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{monte_carlo}
\alias{monte_carlo}
\title{Evaluate model performance with simulated data}
\usage{
monte_carlo(n_reps, real, file = "inst/stuff/Monte_Carlo/simulation.txt", ...)
}
\arguments{
\item{...}{}
}
\description{
Evaluate model performance with simulated data
}
