% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repar_Beta}
\alias{repar_Beta}
\title{Get the parameters of a Beta distribution in terms of its mean and standard deviation}
\usage{
repar_Beta(mu, sd)
}
\arguments{
\item{sd}{}
}
\description{
Get the parameters of a Beta distribution in terms of its mean and standard deviation
}
