% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repar_Gamma}
\alias{repar_Gamma}
\title{Get the parameters of a Gamma distribution in terms of its mean and standard deviation}
\usage{
repar_Gamma(mu, sd)
}
\arguments{
\item{sd}{}
}
\description{
Get the parameters of a Gamma distribution in terms of its mean and standard deviation
}
